+++ src/config.py
@@ -6,7 +6,7 @@ class Config:
     # General config
     batch_size: int = 64
     workers: int = 1
-    dataset_dir: str = '../datasets/zalando-hd-resized'
+    dataset_dir: str = 'datasets/zalando-hd-resized'
     num_epochs: int = 5
     learning_rate: float = 0.0002
     num_classes: int = 10
